Client: Rename aapeli.applet package to aapeli.frame
