+++ client/src/main/java/agolf/GolfGameFrame.java
@@ -3,13 +3,13 @@ package agolf;
 import agolf.game.GameBackgroundCanvas;
 import agolf.game.GamePanel;
 import agolf.lobby.LobbyPanel;
-import com.aapeli.applet.AbstractGameFrame;
 import com.aapeli.client.AutoPopups;
 import com.aapeli.client.BadWordFilter;
 import com.aapeli.client.ImageManager;
 import com.aapeli.client.Parameters;
 import com.aapeli.client.SoundManager;
 import com.aapeli.client.TextManager;
+import com.aapeli.frame.AbstractGameFrame;
 import java.awt.Color;
 import java.awt.Font;
 import java.awt.Image;
