+++ client/src/main/java/com/aapeli/frame/AdCanvas.java
@@ -1,4 +1,4 @@
-package com.aapeli.applet;
+package com.aapeli.frame;
 
 import com.aapeli.client.Parameters;
 import java.awt.Canvas;
