+++ client/src/main/java/com/aapeli/frame/ContentPanel.java
@@ -1,4 +1,4 @@
-package com.aapeli.applet;
+package com.aapeli.frame;
 
 import com.aapeli.client.IPanel;
 import java.awt.Dimension;
